// src/app/Model/UserSetupModel.ts
export class AllUserModel {
    name: String;
    email: String;
    password: String;
    roleId: String;
    role: String;
    shiftID: String;
    condition: String;
    workType: String;
    menu: String
}

export class AllRoleModel {
    condition: string;
    role_id: number;
    role_name: string
}